% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_func.R
\name{exec}
\alias{exec}
\title{Wrapper call function for federated analysis}
\usage{
exec(name, loginFD, logins, func, symbol, ...)
}
\arguments{
\item{name}{A character string naming the function to be called, among federatePCA, federateRCCA, federateComDim, federateSNF.}

\item{loginFD}{Login information of the federated server, where the function \code{name} will be executed. 
For functions using X'X, such as federatePCA, federateRCCA, this should be a server with dsMOprimal installed. 
For functions using XX', such as federateComDim, federateSNF, this should be a server with dsMOdual installed.}

\item{logins}{Login information of data repositories, where dsMOprimal is installed.}

\item{func}{Definition of a function for preparation of raw data matrices.}

\item{symbol}{The symbol provided when calling the function \code{func} for data preparation.}

\item{...}{Other arguments of the function \code{name}, preferably in the same order.}
}
\description{
This wrapper function is used to call all the federated analyse functions provided in the dsMO package suite.
}
\examples{
data(logindata)
data(procFunc)
\donttest{
exec('federatePCA', loginFD=logindata[1,], logins=logindata[1:2,], func=procFunc$SingleOmics, symbol='rawDataX')
exec('federateRCCA', loginFD=logindata[1,], logins=logindata[1:2,], func=procFunc$BiOmics, symbol=c('rawDataX', 'rawDataY'),
      0.001, 0.001, TRUE, 
      dsMO:::.encode.arg(list(nfold = 2, grid1 = seq(0.001, 1, length = 1), grid2 = seq(0.001, 1, length = 1))))
exec('federateComDim', loginFD=logindata[3,], logins=logindata[1:2,], func=procFunc$BiOmics, symbol=c('rawDataX', 'rawDataY'),
      2, 'none', 'uniform')
exec('federateSNF', loginFD=logindata[3,], logins=logindata[1:2,], func=procFunc$BiOmics, symbol=c('rawDataX', 'rawDataY'))
}
}
